Move compaction.status.testutils to upper level
