+++ java/compaction-status-store/src/test/java/sleeper/compaction/status/testutils/CompactionStatusStoreTestUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.status.job.testutils;
+package sleeper.compaction.status.testutils;
 
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
